user_service/requirements.txt
alembic==1.13.1
aniso8601==9.0.1
apispec==6.4.0
asttokens==2.4.1
attrs==23.2.0
blinker==1.7.0
cffi==1.16.0
click==8.1.7
cryptography==42.0.5
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
Flask==3.0.2
Flask-Login==0.6.3
Flask-Migrate==4.0.5
flask-smorest==0.43.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
ipython==8.22.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
Mako==1.3.2
MarkupSafe==2.1.5
marshmallow==3.20.2
matplotlib-inline==0.1.6
mysqlclient==2.2.4
packaging==23.2
parso==0.8.3
pexpect==4.9.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.17.2
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
referencing==0.33.0
rpds-py==0.18.0
six==1.16.0
SQLAlchemy==2.0.27
stack-data==0.6.3
traitlets==5.14.1
typing_extensions==4.9.0
wcwidth==0.2.13
webargs==8.4.0
Werkzeug==3.0.1
